obj/Debug/Package/PackageTmp/requirements.txt
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.3
Django==2.2.16
django-cors-headers==3.5.0
django-enumfield==2.0.2
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pip==20.1.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.1
setuptools==47.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
zipp==3.4.0
awsebcli==3.19.2
botocore==1.19.15
cement==2.8.2
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
future==0.16.0
idna==2.10
jmespath==0.10.0
pathspec==0.5.9
pypiwin32==223
python-dateutil==2.8.1
pywin32==228
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
